Add files via upload
This is my version of the NetworkChuck webdesign template for Wordpress, with minimalistic changes, but more suited to my liking and needs.

It is also worth noting that this is not my code. Subject to changes as I go. Every change has been commented in, but keep in mind that these once again, minimalistic and minor changes for the original.